Added reefing servo controller code and circuit schematics
Added several files related to the reefing servo controller.  Servo Controller Board contains the pcb schematic for the original design of the servo controller, but this was not used.  Servo_controller_uno is the Arduino sketch that was used in the final reefing controller.  This controller was built with an Arduino Uno rather than an Attiny85 as originally planned.